feat: implement PlotCache class for efficient plot caching [full ci]
